move and rename register endpoint
