+++ routes/auth.js
@@ -1,8 +1,6 @@
 const router = require("express").Router();
 const controller = require("../controllers/auth");
 
-router.route("/register").post(controller.newUser);
-
 router.route("/login").post(controller.login);
 
 router.route("/refresh").post(controller.refreshToken);
