+++ routes/users.js
@@ -1,6 +1,6 @@
 const router = require("express").Router();
 const controller = require("../controllers/users");
 
-router.route("/").delete(controller.del);
+router.route("/").post(controller.newUser);
 
 module.exports = router;
