+++ server.js
@@ -22,8 +22,8 @@ app.use(bodyParser.urlencoded({extended: true}));
 app.use(bodyParser.json());
 
 app.use("/api/auth", require("./routes/auth"));
+app.use("/api/users", require("./routes/users"));
 app.use("/api/shows", passport.authenticate("jwt", {session: false}), require("./routes/shows"));
-app.use("/api/users", passport.authenticate("jwt", {session: false}), require("./routes/users"));
 app.use("/api/users/shows", passport.authenticate("jwt", {session: false}), require("./routes/user-shows"));
 
 app.listen(parseInt(process.env.PORT), () => console.log(`Listening on port ${process.env.PORT}`));
